+++ train_mnist.py
@@ -1,13 +1,11 @@
-from pkgutil import get_data
 import torch
 import torch.nn as nn
 import torch.optim as optim
-import torchvision
-from torchvision import transforms
 import matplotlib.pyplot as plt
 from tqdm import tqdm
 
 from src.dataset import load_dataset
+from src.model import SimpleConvNet
 
 
 data_path = '/Users/cesar.salcedo/Documents/datasets'
@@ -21,25 +19,6 @@ data_loader = torch.utils.data.DataLoader(
     shuffle=True,
     num_workers=0)
 
-
-class SimpleConvNet(nn.Module):
-    def __init__(self):
-        super().__init__()
-
-        self.model = nn.Sequential(
-            nn.Conv2d(1, 32, 3, padding=1),
-            nn.MaxPool2d(2),
-            nn.Conv2d(32, 64, 3, padding=1),
-            nn.MaxPool2d(2),
-            nn.Conv2d(64, 128, 3, padding=1),
-            nn.MaxPool2d(2),
-            nn.Conv2d(128, 10, 3, padding=1),
-            nn.MaxPool2d(2),
-        )
-
-    def forward(self, x):
-        y_hat = self.model(x).reshape(-1, 10)
-        return y_hat
 
 model = SimpleConvNet()
 
@@ -57,6 +36,7 @@ for epoch in range(epochs):
         optimizer.step()
 
         losses.append(loss.item())
+
 
 fig = plt.figure()
 plt.plot(losses)
